+++ app/app.py
@@ -4,16 +4,10 @@
 from os import environ as env
 from typing import Final
 
-if "APP_DEBUG_IS_ON" in env:
-    from logging import Logger
+from .helper.configure_logger import debug_on_global, logger_global
 
-    from .helper.configure_logger import configure_logger
+if "APP_SHOW_SYSINFO" in env:
     from .helper.get_system_info import get_system_info
-
-    logger: Logger = configure_logger()
-
-    global debug_on_global
-    debug_on_global: Final = True
 
 from .helper.parse_configs_into_paramdict import get_param_dict
 from .helper.prepare_ml_input import PipelineOutput, prepare_pipeline
@@ -40,11 +34,13 @@ def main(mode: APP_MODES = "train") -> None:
         mode = "train"
 
     if debug_on_global:
-        logger.debug(f"App is running in {mode=}")
-        logger.debug(f"Debug is set to {logger=}")
+
+        logger_global.debug(f"App is running in {mode=}")
+        logger_global.debug(f"Debug is set to {logger_global=}")
+
         if "APP_SHOW_SYSINFO" in env:
             for item in get_system_info():
-                logger.debug(item)
+                logger_global.debug(item)
 
     # pipeline_objects: Pipeline_Output = prepare_pipeline(get_param_dict())
     _: PipelineOutput = prepare_pipeline(get_param_dict())