+++ app/helper/configure_logger.py
@@ -1,13 +1,44 @@
 #!/usr/bin/env python
 """"Load [logger_name] from [root]/[config_path]/[config_fn]"""
 
-from logging import Logger, error, getLogger
+from logging import Logger, debug, error, getLogger
 from logging.config import fileConfig
 from os.path import abspath, dirname, exists, join, split
 
 
-def configure_logger(
-    logger_name: str = "simpleExample",
+def set_global_debug_state_and_logger(debug_on: bool = False):
+    """Toggle global debug state and set global logger"""
+
+    if debug_on is not True:
+        debug_on = False
+
+    debug(f"{debug_on=}")
+
+    _toggle_global_debug_state(debug_on)
+    _set_global_logger(debug_on)
+
+
+def _toggle_global_debug_state(debug_on: bool = False):
+    """Toggle `global debug_on_global` to `debug_on`"""
+
+    global debug_on_global
+    debug_on_global = debug_on
+
+
+def _set_global_logger(debug_on: bool = False):
+    """
+    Set `global logger_global` relative to `debug_on`.
+
+    - `False`: `None`
+    - `True`: `Logger`
+    """
+
+    global logger_global
+    logger_global = _configure_and_get_logger() if debug_on else None
+
+
+def _configure_and_get_logger(
+    logger_name: str = "APP",
     config_fn: str = "logging.conf",
     config_path: str = "config",
 ) -> Logger:
@@ -37,4 +68,5 @@ https://docs.python.org/3/library/logging.config.html#logging.config.fileConfig\
         fileConfig(abs_path)
         return getLogger(logger_name)
     except Exception as e:
+        error(e)
         return e