+++ app/helper/load_configs.py
@@ -1,23 +1,13 @@
 #!/usr/bin/env python
 """Load configuration files"""
 
-from os import environ as env
 from os.path import join
-from typing import Final, Literal, Optional, Union
+from typing import Literal, Optional, Union
 
 from omegaconf import OmegaConf
 
-if "APP_DEBUG_IS_ON" in env:
-    from logging import Logger
-
-    from .configure_logger import configure_logger
-
-    logger: Logger = configure_logger()
-
-    global debug_on_global
-    debug_on_global: Final = True
-
 from .check_sanitize_path import sanitize_path
+from .configure_logger import debug_on_global, logger_global
 
 
 def load_defaults(
@@ -32,7 +22,7 @@ def load_defaults(
     default_configs = _load_config(cfg_defaults_fn, cfg_path)
 
     if debug_on_global:
-        logger.debug(f"{default_configs=}")
+        logger_global.debug(f"{default_configs=}")
 
 
 def get_defaults(key_to_search_and_return: str = "save_dir") -> str:
@@ -51,9 +41,11 @@ def get_config_content(
     if debug_on_global:
         defaults_dbg_msg = f"'{cfg_filename_ex_ext}' found in defaults."
         if _check_or_get_default("config_fn", cfg_filename_ex_ext):
-            logger.debug(defaults_dbg_msg)
+            logger_global.debug(defaults_dbg_msg)
         else:
-            logger.debug(f"No {defaults_dbg_msg} Trying to find the file anyway.")
+            logger_global.debug(
+                f"No {defaults_dbg_msg} Trying to find the file anyway."
+            )
 
     return _load_config(cfg_filename_ex_ext, cfg_path)
 
@@ -68,7 +60,7 @@ def get_keyfile_content(provider: str = "wandb") -> dict:
         keyfile = sanitize_path(keyfiles[provider])
 
         if debug_on_global:
-            logger.debug(f"{keyfile=}")
+            logger_global.debug(f"{keyfile=}")
 
         # TODO refactor to less convoluted call
         return _load_config(keyfile["base"], keyfile["dir"])
@@ -86,7 +78,7 @@ def _load_config(
     """
 
     if debug_on_global:
-        logger.debug(f"Loading {cfg_filename_ex_ext=}")
+        logger_global.debug(f"Loading {cfg_filename_ex_ext=}")
     try:
         # TODO sanitization of yml extension.
         config = OmegaConf.load(join(cfg_path, f"{cfg_filename_ex_ext}.yml"))