+++ app/helper/load_hf_components.py
@@ -5,25 +5,12 @@ Hugging Face caches components into '~/.cache/huggingface'
 """
 # TODO decorator for get_dataset_hf and get_tokenizer_hf
 
-from logging import error
-from os import environ as env
-from typing import Final
-
 from datasets import load_dataset, load_metric
 from datasets.dataset_dict import DatasetDict
 from transformers import AutoModel, AutoTokenizer
 
-if "APP_DEBUG_IS_ON" in env:
-    from logging import Logger
-
-    from .configure_logger import configure_logger
-
-    logger: Logger = configure_logger()
-
-    global debug_on_global
-    debug_on_global: Final = True
-
 from .check_sanitize_path import check_and_create_path
+from .configure_logger import debug_on_global, logger_global
 
 
 def get_dataset_hf(
@@ -56,23 +43,25 @@ https://huggingface.co/docs/datasets/loading#local-and-remote-files\
         msg_config = f"{configuration=} from " if configuration else ""
         msg_ds_full = f"{msg_config}{dataset_name=}"
         if path_exists:
-            logger.debug(f"Loading local copy of {msg_ds_full} from {save_path=}")
+            logger_global.debug(
+                f"Loading local copy of {msg_ds_full} from {save_path=}"
+            )
         else:
-            logger.debug(f"Downloading {msg_ds_full}")
+            logger_global.debug(f"Downloading {msg_ds_full}")
 
     try:
         dataset = load_dataset(**ds_load_params)
         if not path_exists:
             if debug_on_global:
-                logger.debug(f"Saving dataset to {save_path=}")
+                logger_global.debug(f"Saving dataset to {save_path=}")
             dataset.save_to_disk(save_path)
     except Exception as e:
-        error(e)
+        logger_global.error(e)
         return e
 
     if debug_on_global:
         ds_train_slice = dataset["train"][0]
-        logger.debug(f"{ds_train_slice=}")
+        logger_global.debug(f"{ds_train_slice=}")
 
     return dataset
 
@@ -100,24 +89,24 @@ model_doc/auto#transformers.AutoTokenizer\
     if debug_on_global:
         msg_tok = f"tokenizer for {model_name=}"
         if path_exists:
-            logger.debug(f"Loading local copy of {msg_tok} from {save_path=}")
+            logger_global.debug(f"Loading local copy of {msg_tok} from {save_path=}")
         else:
-            logger.debug(f"Downloading {msg_tok}")
+            logger_global.debug(f"Downloading {msg_tok}")
 
     try:
         tokenizer = AutoTokenizer.from_pretrained(**tokenizer_load_params)
         if not path_exists:
             if debug_on_global:
-                logger.debug(f"Saving tokenizer to {save_path=}")
+                logger_global.debug(f"Saving tokenizer to {save_path=}")
             tokenizer.save_pretrained(save_path)
     except Exception as e:
-        error(e)
+        logger_global.error(e)
         return e
 
     if debug_on_global:
         tok_msg = "This is a test sentence for the loaded tokenizer."
         tok_res = tokenizer.encode(tok_msg)
-        logger.debug(f"Tokenizing '{tok_msg}': {tok_res}")
+        logger_global.debug(f"Tokenizing '{tok_msg}': {tok_res}")
 
     return tokenizer
 
@@ -159,7 +148,7 @@ def get_model_hf(model_full_name: str, num_labels: int) -> AutoModel:
     #     print(red, e)
 
     if debug_on_global:
-        logger.debug(f"Downloading {model_full_name=} with {num_labels=}")
+        logger_global.debug(f"Downloading {model_full_name=} with {num_labels=}")
 
     return AutoModel.from_pretrained(model_full_name, num_labels=num_labels)
 
@@ -168,22 +157,24 @@ def get_metrics_to_load_objects_hf(metrics_to_load: list) -> list[dict]:
     """Downloads Hugging Face Metrics Builder Scripts"""
 
     # TODO metrics save and load locally possible ?
-    # TODO error handling, what about empty metrics?
+    # TODO logger_global.error( handling, what about empty metrics?
 
     if debug_on_global:
-        logger.debug(f"Loading HF Metrics Builder Scripts for {metrics_to_load!r}")
+        logger_global.debug(
+            f"Loading HF Metrics Builder Scripts for {metrics_to_load!r}"
+        )
 
     metrics_loaded = []
 
     for met in metrics_to_load:
 
         if debug_on_global:
-            logger.debug(f"Trying to load {met!r}")
+            logger_global.debug(f"Trying to load {met!r}")
 
         try:
             metrics_loaded.append(load_metric(met))
         except Exception as e:
-            # TODO handle error while loading metrics from HF
-            error(e)
+            # TODO handle logger_global.error( while loading metrics from HF
+            logger_global.error(e)
 
     return metrics_loaded